+++ src/relstorage/adapters/postgresql/connmanager.py
@@ -91,10 +91,6 @@ class Psycopg2ConnectionManager(AbstractConnectionManager):
                 raise
 
     def _do_open_for_load(self):
-        """Open and initialize a connection for loading objects.
-
-        Returns (conn, cursor).
-        """
         # XXX: SERIALIZABLE isn't allowed on streaming replicas
         # (https://www.enterprisedb.com/blog/serializable-postgresql-11-and-beyond)
         # Do we really need SERIALIZABLE? Wouldn't REPEATABLE READ be
@@ -103,22 +99,13 @@ class Psycopg2ConnectionManager(AbstractConnectionManager):
         # Set the transaction to READ ONLY mode. This lets
         # transactions (especially SERIALIZABLE) elide some locks.
 
-        # TODO: Enable deferrable transactions if we stay in serializable, read only
-        # mode. This should generally be faster, as the *only* serializable transactions
-        # we have should be READ ONLY.
+        # TODO: Enable deferrable transactions if we stay in
+        # serializable, read only mode. This should generally be
+        # faster, as the *only* serializable transactions we have
+        # should be READ ONLY.
         return self.open(self.isolation_serializable,
                          read_only=True,
                          replica_selector=self.ro_replica_selector)
-
-    # TODO: Define _do_open_for_store to change the default isolation
-    # level from READ COMMITTED to REPEATABLE READ? READ COMMITTED
-    # takes a MVCC snapshot for every individual *statement* in the
-    # transaction; REPEATABLE READ takes a snapshot for the whole
-    # duration of the transaction, beginning at the first
-    # SELECT/UPDATE/INSERT statement. We could get by with fewer locks
-    # that way (see locker.py).
-    # def _do_open_for_store(self):
-    #    return self.open('REPEATABLE READ')
 
     def open_for_pre_pack(self):
         return self.open(self.isolation_read_committed)